more refactoring and cleaning
